+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fixed bug in concurrency calculation that allowed too many active requests
+- Fixed bug in detecting concurrency limit exceeded Qualys API response
 
 ## 5.0.3 - 2020-01-04